Added a ProgressDetails component to handle the details of earned progress as modal screen. EW-242
Initial modal screen to show progress details. To be used in ProgressCard and ProgressListItem components to show progress detail.

Upgraded Paragon library to 4.1.4 release to allow for Button component children prop information. This allowed embedding a image within the button element.

Todo: Need to add in recipient and download information.