Added environment values used within the `badges-frontend` application. EW-237
Most of these environment variables were used for the `@edx/frontend-auth` application to talk back with the LMS ReST APIs. At this time we're not using that but instead using NGINX to proxy back our requests to the LMS like `@edx/studio-frontend` is doing.

The image path for `/src/images/badgr-logo.svg` was missing so setting the `BASE_URL` variable will help with that.

For the most part these environment settings will be not set for the Hawthorn release. `MOCK_LMS_API` is used to pull in mock data from `/src/data/services/LmsApiService.js` to simulate a call to the LMS ReST API but should be set to `false` in production mode.